refactor to support latest etcd
Made changes necessary to support new signature for etcd.Client.Get. Was
Get(key string). Now Get(key string, sort, recursive bool)

Made changes to support new structure of etcd.Response.

Refactored etcdutil.etcdWalk to use pointers to nodes to walk the tree,
eliminating the need to query etcd for each child node. This work also exposed
an infinite loop in the previous implementation of etcdWalk when one of
the keys in the toml configuration was a node whose Dir property was
true.