Always trim prefix from keys
It is common for prefixed keys to be stored in etcd with a
leading "/". When retrieving keys from etcd, the leading "/"
is preserved, this is not the case for consul, which means
the prefix doesn't always get stripped off.

Ensure the prefix is always removed by trimming the leading "/"
from all keys and the user defined prefix before trimming the
prefix from the key.