Add new flag -keep-stage-file
Currently it's impossible to debug issues when the check command
fails because the staged file being checked is deleted.

Add a flag, -keep-stage-file, to keep the stage files around after
confd exits and log the full path to the staged file.

The confd logs now print the following log message if the
-keep-stage-file flag is used:

    INFO Keeping staged file: /tmp/.backend-upstream.conf018905335

Fixes #97