add Makefile for building confd from source
This aids users with interacting with the build process behind confd.
With a Makefile, users just need to run `make` to build confd, then
`make install` to install it to /usr/local/bin/confd. This path can be
overridden through make:

    $ make prefix=$HOME install

Since TravisCI overrides the script to `make` when a makefile is
present, I also had to modify the travis.yml file to test the project.