add --log-level flag
This refactors the log package to use logrus in the background, while
also introducing the `--log-level` flag. This also deprecates flags like
`--quiet`, `--verbose` and `--debug` as all three are covered in one
flag. If a unrecognized log level is introduced, an error is displayed
and `confd` calls `log.Fatal()` with a proper response.

The justification behind switching to logrus allows us in the future to
take advantage of the many features which exist in logrus, such as their
[hook infrastructure](https://github.com/Sirupsen/logrus#hooks).