Prefix set on cli overrides configs and used with watches
confd is currently appending the prefix set on the cli with the
prefix set in template resource configs. This is incorrect. The
cli should override the configuration file.

confd is also not using the prefix supplied by the user and always
watches using the "/" prefix.

Fix these issues by giving priority to the prefix set on the cli
and using the current prefix for watches.