Don't pass empty prefix to zookeeper nodeWalk
https://github.com/kelseyhightower/confd/issues/656

The zookeeper backend attempts to call nodeWalk with an empty prefix.  Newer
versions of https://github.com/samuel/go-zookeeper/ will validate keys passed
as input parameters and return an error if an empty string is passed.  This
change modifies the logic slightly to handle the case where we want to walk the
nodes of the parent "/" key.