Google Metadata: Make tests happy
Instead of returning raw JSON dumps of directories, actually return the
correct key/value maps.  We only return JSON for things we don't
recognize.  instance/tags (which is a list) is the one known attribute
in this case.

You can store raw JSON in an attribute, but that will be returned by the
metadata API as a string so should provide the expected key mapping to a
JSON encoded value the user originally stored in the metadata key.