Upgrade Go to 1.17 and Refactor the project layout
Have been working to resolve the security vulnerability identified
by dependabot. It seemed the only way to accomplish that was to
update to the latest version of Go. This also seemed like a good
time to adopt a more conventional project layout using the
golang-standard/project-layout as reference.

Finally, this version will drop the support for Rancher metadata as
those codebases have been archived. It's unclear what, if any
metadata is available in the 2.X versions of Rancher.